Trait vmm_vhost::connection::Endpoint

pub trait Endpoint<R: Req>: Sized {
    type Listener: Listener;
    fn from_connection(sock: <Self::Listener as Listener>::Connection) -> Self;

    fn connect<P: AsRef<Path>>(path: P) -> Result<Self>;

    fn send_iovec(
        &mut self, 
        iovs: &[IoSlice<'_>], 
        fds: Option<&[RawFd]>
    ) -> Result<usize>;

    fn recv_into_bufs(
        &mut self, 
        bufs: &mut [IoSliceMut<'_>], 
        allow_fd: bool
    ) -> Result<(usize, Option<Vec<File>>)>;
}

Abstracts a vhost-user connection and related operations.

Associated Types

type Listener: Listener

Type of an object that Endpoint is created from.

Required methods

fn from_connection(sock: <Self::Listener as Listener>::Connection) -> Self

Create an endpoint from a stream object.

fn connect<P: AsRef<Path>>(path: P) -> Result<Self>

Create a new stream by connecting to server at str.

fn send_iovec(
&mut self,
iovs: &[IoSlice<'_>],
fds: Option<&[RawFd]>
) -> Result<usize>

Sends bytes from scatter-gather vectors with optional attached file descriptors.

Return:

number of bytes sent on success

fn recv_into_bufs(
&mut self,
bufs: &mut [IoSliceMut<'_>],
allow_fd: bool
) -> Result<(usize, Option<Vec<File>>)>

Reads bytes into the given scatter/gather vectors with optional attached file.

Arguements

bufs - A slice of buffers to store received data.

allow_fd - Indicates whether we can receive FDs.

Return:

(number of bytes received, [received files]) on success