Trait vmm_vhost::connection::Listener

pub trait Listener: Sized {
    type Connection;
    type Endpoint;
    fn accept(&mut self) -> Result<Option<Self::Endpoint>>;

    fn set_nonblocking(&self, block: bool) -> Result<()>;
}

Listener for accepting connections.

Associated Types

type Connection

Type of an object created when a connection is accepted.

type Endpoint

Type of endpoint created when a connection is accepted.

Required methods

fn accept(&mut self) -> Result<Option<Self::Endpoint>>

Accept an incoming connection.

fn set_nonblocking(&self, block: bool) -> Result<()>

Change blocking status on the listener.