Trait vmm_vhost::connection::vfio::Device

pub trait Device: Send {
    fn event(&self) -> &Event;

    fn start(&mut self) -> Result<(), Error>;

    fn send_bufs(
        &mut self, 
        iovs: &[IoSlice<'_>], 
        fds: Option<&[RawFd]>
    ) -> Result<usize, Error>;

    fn recv_into_bufs(
        &mut self, 
        iovs: &mut [IoSliceMut<'_>]
    ) -> Result<usize, RecvIntoBufsError>;

    fn create_slave_request_endpoint(
        &mut self
    ) -> Result<Box<dyn EndpointTrait<SlaveReq>>, Error>;
}

VFIO device which can be used as virtio-vhost-user device backend.

Required methods

fn event(&self) -> &Event

This event must be read before handle_request() is called.

fn start(&mut self) -> Result<(), Error>

Starts VFIO device.

fn send_bufs(
&mut self,
iovs: &[IoSlice<'_>],
fds: Option<&[RawFd]>
) -> Result<usize, Error>

Sends data in the given slice of slices.

fn recv_into_bufs(
&mut self,
iovs: &mut [IoSliceMut<'_>]
) -> Result<usize, RecvIntoBufsError>

Receives data into the given slice of slices and returns the size of the received data.

fn create_slave_request_endpoint(
&mut self
) -> Result<Box<dyn EndpointTrait<SlaveReq>>, Error>

Constructs the slave request endpoint for the endpoint backed by this device.