Enum vmm_vhost::Error

pub enum Error {

    ClientExit,
    Disconnect,
    FeatureMismatch,
    IncorrectFds,
    InvalidMessage,
    InvalidOperation,
    InvalidParam,
    MasterInternalError,
    OversizedMsg,
    PartialMessage,
    RecvBufferTooSmall {
        got: usize,
        want: usize,
    },
    ReqHandlerError(IOError),
    SlaveInternalError,
    SocketBroken(Error),
    SocketConnect(Error),
    SocketError(Error),
    SocketRetry(Error),
    TubeError(TubeError),
    VfioDeviceError(Error),

}

Errors for vhost-user operations

Variants

ClientExit

client exited properly.

Disconnect

client disconnected. If connection is closed properly, use ClientExit instead.

FeatureMismatch

Virtio/protocol features mismatch.

IncorrectFds

Fd array in question is too big or too small

InvalidMessage

Invalid message format, flag or content.

InvalidOperation

Unsupported operations due to that the protocol feature hasn’t been negotiated.

InvalidParam

Invalid parameters.

MasterInternalError

Failure from the master side.

OversizedMsg

Message is too large

PartialMessage

Only part of a message have been sent or received successfully

RecvBufferTooSmall

Fields

got: usize

The size of the buffer received.

want: usize

The expected size of the buffer.

Provided recv buffer was too small, and data was dropped.

ReqHandlerError(IOError)

Error from request handler

SlaveInternalError

Failure from the slave side.

SocketBroken(Error)

The socket is broken or has been closed.

SocketConnect(Error)

Can’t connect to peer.

SocketError(Error)

Generic socket errors.

SocketRetry(Error)

Should retry the socket operation again.

TubeError(TubeError)

Error from tx/rx on a Tube.

VfioDeviceError(Error)

Error from VFIO device.