Struct vmm_vhost::master::Master

pub struct Master<E: Endpoint<MasterReq>> {
    node: Arc<Mutex<MasterInternal<E>>>,
}

Struct for the vhost-user master endpoint.

Fields
node: Arc<Mutex<MasterInternal<E>>>
Implementations

impl<E: Endpoint<MasterReq> + From<SystemStream>> Master<E>

pub fn from_stream(sock: SystemStream, max_queue_num: u64) -> Self

Create a new instance from a Unix stream socket.

impl<E: Endpoint<MasterReq>> Master<E>

fn new(ep: E, max_queue_num: u64) -> Self

Create a new instance.

fn node(&self) -> MutexGuard<'_, MasterInternal<E>>

pub fn connect<P: AsRef<Path>>(path: P, max_queue_num: u64) -> Result<Self>

Create a new vhost-user master endpoint.

Will retry as the backend may not be ready to accept the connection.

Arguments

path - path of Unix domain socket listener to connect to

pub fn set_hdr_flags(&self, flags: VhostUserHeaderFlag)

Set the header flags that should be applied to all following messages.