Struct vmm_vhost::Master

pub struct Master<E: Endpoint<MasterReq>> { /* fields omitted */ }

Struct for the vhost-user master endpoint.

Implementations

impl<E: Endpoint<MasterReq> + From<UnixStream>> Master<E>

pub fn from_stream(sock: UnixStream, max_queue_num: u64) -> Self

Create a new instance from a Unix stream socket.

impl<E: Endpoint<MasterReq>> Master<E>

pub fn connect<P: AsRef<Path>>(path: P, max_queue_num: u64) -> Result<Self>

Create a new vhost-user master endpoint.

Will retry as the backend may not be ready to accept the connection.

Arguments

path - path of Unix domain socket listener to connect to

pub fn set_hdr_flags(&self, flags: VhostUserHeaderFlag)

Set the header flags that should be applied to all following messages.