Struct vmm_vhost::Slave

pub struct Slave {
    node: Arc<Mutex<SlaveInternal>>,
}

Request proxy to send slave requests to the master through the slave communication channel.

The Slave acts as a message proxy to forward slave requests to the master through the vhost-user slave communication channel. The forwarded messages will be handled by the MasterReqHandler server.

Fields
node: Arc<Mutex<SlaveInternal>>
Implementations

impl Slave

pub fn new(ep: Box<dyn Endpoint<SlaveReq>>) -> Self

Constructs a new slave proxy from the given endpoint.

fn node(&self) -> MutexGuard<'_, SlaveInternal>

fn send_message<T>(
&self,
request: SlaveReq,
msg: &T,
fds: Option<&[RawDescriptor]>
) -> Result<u64> where
T: DataInit,

pub fn from_stream(sock: SystemStream) -> Self

Create a new instance from a SystemStream object.

pub fn set_reply_ack_flag(&self, enable: bool)

Set the negotiation state of the VHOST_USER_PROTOCOL_F_REPLY_ACK protocol feature.

When the VHOST_USER_PROTOCOL_F_REPLY_ACK protocol feature has been negotiated, the “REPLY_ACK” flag will be set in the message header for every slave to master request message.

pub fn set_failed(&self, error: i32)

Mark endpoint as failed with specified error code.