Struct vmm_vhost::SlaveFsCacheReq

pub struct SlaveFsCacheReq {
    node: Arc<Mutex<SlaveFsCacheReqInternal>>,
}

Request proxy to send vhost-user-fs slave requests to the master through the slave communication channel.

The SlaveFsCacheReq acts as a message proxy to forward vhost-user-fs slave requests to the master through the vhost-user slave communication channel. The forwarded messages will be handled by the MasterReqHandler server.

Fields
node: Arc<Mutex<SlaveFsCacheReqInternal>>
Implementations

impl SlaveFsCacheReq

fn new(ep: SocketEndpoint<SlaveReq>) -> Self

fn node(&self) -> MutexGuard<'_, SlaveFsCacheReqInternal>

fn send_message(
&self,
request: SlaveReq,
fs: &VhostUserFSSlaveMsg,
fds: Option<&[RawDescriptor]>
) -> Result<u64>

pub fn from_stream(sock: SystemStream) -> Self

Create a new instance from a SystemStream object.

pub fn set_reply_ack_flag(&self, enable: bool)

Set the negotiation state of the VHOST_USER_PROTOCOL_F_REPLY_ACK protocol feature.

When the VHOST_USER_PROTOCOL_F_REPLY_ACK protocol feature has been negotiated, the “REPLY_ACK” flag will be set in the message header for every slave to master request message.

pub fn set_failed(&self, error: i32)

Mark endpoint as failed with specified error code.