Struct vmm_vhost::SlaveReqHandler

pub struct SlaveReqHandler<S: VhostUserSlaveReqHandler, E: Endpoint<MasterReq>> {
    slave_req_helper: SlaveReqHelper<E>,
    backend: S,
    virtio_features: u64,
    acked_virtio_features: u64,
    protocol_features: VhostUserProtocolFeatures,
    acked_protocol_features: u64,
    error: Option<i32>,
}

Server to handle service requests from masters from the master communication channel.

The SlaveReqHandler acts as a server on the slave side, to handle service requests from masters on the master communication channel. It’s actually a proxy invoking the registered handler implementing VhostUserSlaveReqHandler to do the real work.

The lifetime of the SlaveReqHandler object should be the same as the underline Unix Domain Socket, so it gets simpler to recover from disconnect.

Fields
slave_req_helper: SlaveReqHelper<E>backend: Svirtio_features: u64acked_virtio_features: u64protocol_features: VhostUserProtocolFeaturesacked_protocol_features: u64error: Option<i32>
Implementations

impl<S: VhostUserSlaveReqHandler> SlaveReqHandler<S, SocketEndpoint<MasterReq>>

pub fn from_stream(socket: SystemStream, backend: S) -> Self

Create a vhost-user slave endpoint from a connected socket.

impl<S: VhostUserSlaveReqHandler, E: Endpoint<MasterReq>> SlaveReqHandler<S, E>

pub fn new(endpoint: E, backend: S) -> Self

Create a vhost-user slave endpoint.

pub fn connect(path: &str, backend: S) -> Result<Self>

Create a new vhost-user slave endpoint.

Arguments

path - path of Unix domain socket listener to connect to

backend - handler for requests from the master to the slave

pub fn set_failed(&mut self, error: i32)

Mark endpoint as failed with specified error code.

pub fn handle_request(&mut self) -> Result<()>

Main entrance to request from the communication channel.

Receive and handle one incoming request message from the vmm. The caller needs to:

serialize calls to this function

decide what to do when error happens

optional recover from failure

Return:

Ok(()): one request was successfully handled.

Err(ClientExit): the vmm closed the connection properly. This isn’t an actual failure.

Err(Disconnect): the connection was closed unexpectedly.

Err(InvalidMessage): the vmm sent a illegal message.

other errors: failed to handle a request.

fn set_mem_table(
&mut self,
hdr: &VhostUserMsgHeader<MasterReq>,
size: usize,
buf: &[u8],
files: Option<Vec<File>>
) -> Result<()>

fn get_config(
&mut self,
hdr: &VhostUserMsgHeader<MasterReq>,
buf: &[u8]
) -> Result<()>

fn set_config(&mut self, size: usize, buf: &[u8]) -> Result<()>

fn set_slave_req_fd(&mut self, files: Option<Vec<File>>) -> Result<()>

fn handle_vring_fd_request(
&mut self,
buf: &[u8],
files: Option<Vec<File>>
) -> Result<(u8, Option<File>)>

fn check_state(&self) -> Result<()>

fn check_request_size(
&self,
hdr: &VhostUserMsgHeader<MasterReq>,
size: usize,
expected: usize
) -> Result<()>

fn check_attached_files(
&self,
hdr: &VhostUserMsgHeader<MasterReq>,
files: &Option<Vec<File>>
) -> Result<()>

fn extract_request_body<T: Sized + DataInit + VhostUserMsgValidator>(
&self,
hdr: &VhostUserMsgHeader<MasterReq>,
size: usize,
buf: &[u8]
) -> Result<T>

fn update_reply_ack_flag(&mut self)