Trait vmm_vhost::VhostBackendMut

pub trait VhostBackendMut: Sized {

    fn get_features(&mut self) -> Result<u64>;

    fn set_features(&mut self, features: u64) -> Result<()>;

    fn set_owner(&mut self) -> Result<()>;

    fn reset_owner(&mut self) -> Result<()>;

    fn set_mem_table(
        &mut self, 
        regions: &[VhostUserMemoryRegionInfo]
    ) -> Result<()>;

    fn set_log_base(
        &mut self, 
        base: u64, 
        fd: Option<RawDescriptor>
    ) -> Result<()>;

    fn set_log_fd(&mut self, fd: RawDescriptor) -> Result<()>;

    fn set_vring_num(&mut self, queue_index: usize, num: u16) -> Result<()>;

    fn set_vring_addr(
        &mut self, 
        queue_index: usize, 
        config_data: &VringConfigData
    ) -> Result<()>;

    fn set_vring_base(&mut self, queue_index: usize, base: u16) -> Result<()>;

    fn get_vring_base(&mut self, queue_index: usize) -> Result<u32>;

    fn set_vring_call(
        &mut self, 
        queue_index: usize, 
        event: &Event
    ) -> Result<()>;

    fn set_vring_kick(
        &mut self, 
        queue_index: usize, 
        event: &Event
    ) -> Result<()>;

    fn set_vring_err(&mut self, queue_index: usize, event: &Event) -> Result<()>;

}

An interface for setting up vhost-based backend drivers.

Vhost devices are subset of virtio devices, which improve virtio device’s performance by delegating data plane operations to dedicated IO service processes. Vhost devices use the same virtqueue layout as virtio devices to allow vhost devices to be mapped directly to virtio devices.

The purpose of vhost is to implement a subset of a virtio device’s functionality outside the VMM process. Typically fast paths for IO operations are delegated to the dedicated IO service processes, and slow path for device configuration are still handled by the VMM process. It may also be used to control access permissions of virtio backend devices.

Required methods

fn get_features(&mut self) -> Result<u64>

Get a bitmask of supported virtio/vhost features.

fn set_features(&mut self, features: u64) -> Result<()>

Inform the vhost subsystem which features to enable. This should be a subset of supported features from get_features().

Arguments

features - Bitmask of features to set.

fn set_owner(&mut self) -> Result<()>

Set the current process as the owner of the vhost backend. This must be run before any other vhost commands.

fn reset_owner(&mut self) -> Result<()>

Used to be sent to request disabling all rings This is no longer used.

fn set_mem_table(&mut self, regions: &[VhostUserMemoryRegionInfo]) -> Result<()>

Set the guest memory mappings for vhost to use.

fn set_log_base(&mut self, base: u64, fd: Option<RawDescriptor>) -> Result<()>

Set base address for page modification logging.

fn set_log_fd(&mut self, fd: RawDescriptor) -> Result<()>

Specify an event file descriptor to signal on log write.

fn set_vring_num(&mut self, queue_index: usize, num: u16) -> Result<()>

Set the number of descriptors in the vring.

Arguments

queue_index - Index of the queue to set descriptor count for.

num - Number of descriptors in the queue.

fn set_vring_addr(
&mut self,
queue_index: usize,
config_data: &VringConfigData
) -> Result<()>

Set the addresses for a given vring.

Arguments

queue_index - Index of the queue to set addresses for.

config_data - Configuration data for a vring.

fn set_vring_base(&mut self, queue_index: usize, base: u16) -> Result<()>

Set the first index to look for available descriptors.

Arguments

queue_index - Index of the queue to modify.

num - Index where available descriptors start.

fn get_vring_base(&mut self, queue_index: usize) -> Result<u32>

Get the available vring base offset.

fn set_vring_call(&mut self, queue_index: usize, event: &Event) -> Result<()>

Set the event to trigger when buffers have been used by the host.

Arguments

queue_index - Index of the queue to modify.

event - Event to trigger.

fn set_vring_kick(&mut self, queue_index: usize, event: &Event) -> Result<()>

Set the event that will be signaled by the guest when buffers are available for the host to process.

Arguments

queue_index - Index of the queue to modify.

event - Event that will be signaled from guest.

fn set_vring_err(&mut self, queue_index: usize, event: &Event) -> Result<()>

Set the event that will be signaled by the guest when error happens.

Arguments

queue_index - Index of the queue to modify.

event - Event that will be signaled from guest.