Trait vmm_vhost::VhostUserMaster

pub trait VhostUserMaster: VhostBackend {
    fn get_protocol_features(&mut self) -> Result<VhostUserProtocolFeatures>;

    fn set_protocol_features(
        &mut self, 
        features: VhostUserProtocolFeatures
    ) -> Result<()>;

    fn get_queue_num(&mut self) -> Result<u64>;

    fn set_vring_enable(
        &mut self, 
        queue_index: usize, 
        enable: bool
    ) -> Result<()>;

    fn get_config(
        &mut self, 
        offset: u32, 
        size: u32, 
        flags: VhostUserConfigFlags, 
        buf: &[u8]
    ) -> Result<(VhostUserConfig, VhostUserConfigPayload)>;

    fn set_config(
        &mut self, 
        offset: u32, 
        flags: VhostUserConfigFlags, 
        buf: &[u8]
    ) -> Result<()>;

    fn set_slave_request_fd(&mut self, fd: &dyn AsRawFd) -> Result<()>;

    fn get_inflight_fd(
        &mut self, 
        inflight: &VhostUserInflight
    ) -> Result<(VhostUserInflight, File)>;

    fn set_inflight_fd(
        &mut self, 
        inflight: &VhostUserInflight, 
        fd: RawFd
    ) -> Result<()>;

    fn get_max_mem_slots(&mut self) -> Result<u64>;

    fn add_mem_region(
        &mut self, 
        region: &VhostUserMemoryRegionInfo
    ) -> Result<()>;

    fn remove_mem_region(
        &mut self, 
        region: &VhostUserMemoryRegionInfo
    ) -> Result<()>;
}

Trait for vhost-user master to provide extra methods not covered by the VhostBackend yet.

Required methods

fn get_protocol_features(&mut self) -> Result<VhostUserProtocolFeatures>

Get the protocol feature bitmask from the underlying vhost implementation.

fn set_protocol_features(
&mut self,
features: VhostUserProtocolFeatures
) -> Result<()>

Enable protocol features in the underlying vhost implementation.

fn get_queue_num(&mut self) -> Result<u64>

Query how many queues the backend supports.

fn set_vring_enable(&mut self, queue_index: usize, enable: bool) -> Result<()>

Signal slave to enable or disable corresponding vring.

Slave must not pass data to/from the backend until ring is enabled by VHOST_USER_SET_VRING_ENABLE with parameter 1, or after it has been disabled by VHOST_USER_SET_VRING_ENABLE with parameter 0.

fn get_config(
&mut self,
offset: u32,
size: u32,
flags: VhostUserConfigFlags,
buf: &[u8]
) -> Result<(VhostUserConfig, VhostUserConfigPayload)>

Fetch the contents of the virtio device configuration space.

fn set_config(
&mut self,
offset: u32,
flags: VhostUserConfigFlags,
buf: &[u8]
) -> Result<()>

Change the virtio device configuration space. It also can be used for live migration on the destination host to set readonly configuration space fields.

fn set_slave_request_fd(&mut self, fd: &dyn AsRawFd) -> Result<()>

Setup slave communication channel.

fn get_inflight_fd(
&mut self,
inflight: &VhostUserInflight
) -> Result<(VhostUserInflight, File)>

Retrieve shared buffer for inflight I/O tracking.

fn set_inflight_fd(
&mut self,
inflight: &VhostUserInflight,
fd: RawFd
) -> Result<()>

Set shared buffer for inflight I/O tracking.

fn get_max_mem_slots(&mut self) -> Result<u64>

Query the maximum amount of memory slots supported by the backend.

fn add_mem_region(&mut self, region: &VhostUserMemoryRegionInfo) -> Result<()>

Add a new guest memory mapping for vhost to use.

fn remove_mem_region(
&mut self,
region: &VhostUserMemoryRegionInfo
) -> Result<()>

Remove a guest memory mapping from vhost.