Trait vmm_vhost::VhostUserMasterReqHandler

pub trait VhostUserMasterReqHandler {
    fn handle_config_change(&self) -> HandlerResult<u64> { ... }

    fn fs_slave_map(
        &self, 
        _fs: &VhostUserFSSlaveMsg, 
        _fd: &dyn AsRawFd
    ) -> HandlerResult<u64> { ... }

    fn fs_slave_unmap(&self, _fs: &VhostUserFSSlaveMsg) -> HandlerResult<u64> { ... }

    fn fs_slave_sync(&self, _fs: &VhostUserFSSlaveMsg) -> HandlerResult<u64> { ... }

    fn fs_slave_io(
        &self, 
        _fs: &VhostUserFSSlaveMsg, 
        _fd: &dyn AsRawFd
    ) -> HandlerResult<u64> { ... }
}

Define services provided by masters for the slave communication channel.

The vhost-user specification defines a slave communication channel, by which slaves could request services from masters. The VhostUserMasterReqHandler trait defines services provided by masters, and it’s used both on the master side and slave side.

on the slave side, a stub forwarder implementing VhostUserMasterReqHandler will proxy service requests to masters. The SlaveFsCacheReq is an example stub forwarder.

on the master side, the MasterReqHandler will forward service requests to a handler implementing VhostUserMasterReqHandler.

The VhostUserMasterReqHandler trait is design with interior mutability to improve performance for multi-threading.

Provided methods

fn handle_config_change(&self) -> HandlerResult<u64>

Handle device configuration change notifications.

fn fs_slave_map(
&self,
_fs: &VhostUserFSSlaveMsg,
_fd: &dyn AsRawFd
) -> HandlerResult<u64>

Handle virtio-fs map file requests.

fn fs_slave_unmap(&self, _fs: &VhostUserFSSlaveMsg) -> HandlerResult<u64>

Handle virtio-fs unmap file requests.

fn fs_slave_sync(&self, _fs: &VhostUserFSSlaveMsg) -> HandlerResult<u64>

Handle virtio-fs sync file requests.

fn fs_slave_io(
&self,
_fs: &VhostUserFSSlaveMsg,
_fd: &dyn AsRawFd
) -> HandlerResult<u64>

Handle virtio-fs file IO requests.