Enum vmm_vhost::vhost_user::message::MasterReq

#[repr(u32)]pub enum MasterReq {
    NOOP,
    GET_FEATURES,
    SET_FEATURES,
    SET_OWNER,
    RESET_OWNER,
    SET_MEM_TABLE,
    SET_LOG_BASE,
    SET_LOG_FD,
    SET_VRING_NUM,
    SET_VRING_ADDR,
    SET_VRING_BASE,
    GET_VRING_BASE,
    SET_VRING_KICK,
    SET_VRING_CALL,
    SET_VRING_ERR,
    GET_PROTOCOL_FEATURES,
    SET_PROTOCOL_FEATURES,
    GET_QUEUE_NUM,
    SET_VRING_ENABLE,
    SEND_RARP,
    NET_SET_MTU,
    SET_SLAVE_REQ_FD,
    IOTLB_MSG,
    SET_VRING_ENDIAN,
    GET_CONFIG,
    SET_CONFIG,
    CREATE_CRYPTO_SESSION,
    CLOSE_CRYPTO_SESSION,
    POSTCOPY_ADVISE,
    POSTCOPY_LISTEN,
    POSTCOPY_END,
    GET_INFLIGHT_FD,
    SET_INFLIGHT_FD,
    GPU_SET_SOCKET,
    RESET_DEVICE,
    VRING_KICK,
    GET_MAX_MEM_SLOTS,
    ADD_MEM_REG,
    REM_MEM_REG,
    SET_STATUS,
    GET_STATUS,
    MAX_CMD,
}

Type of requests sending from masters to slaves.

Variants

NOOP

Null operation.

GET_FEATURES

Get from the underlying vhost implementation the features bit mask.

SET_FEATURES

Enable features in the underlying vhost implementation using a bit mask.

SET_OWNER

Set the current Master as an owner of the session.

RESET_OWNER

No longer used.

SET_MEM_TABLE

Set the memory map regions on the slave so it can translate the vring addresses.

SET_LOG_BASE

Set logging shared memory space.

SET_LOG_FD

Set the logging file descriptor, which is passed as ancillary data.

SET_VRING_NUM

Set the size of the queue.

SET_VRING_ADDR

Set the addresses of the different aspects of the vring.

SET_VRING_BASE

Set the base offset in the available vring.

GET_VRING_BASE

Get the available vring base offset.

SET_VRING_KICK

Set the event file descriptor for adding buffers to the vring.

SET_VRING_CALL

Set the event file descriptor to signal when buffers are used.

SET_VRING_ERR

Set the event file descriptor to signal when error occurs.

GET_PROTOCOL_FEATURES

Get the protocol feature bit mask from the underlying vhost implementation.

SET_PROTOCOL_FEATURES

Enable protocol features in the underlying vhost implementation.

GET_QUEUE_NUM

Query how many queues the backend supports.

SET_VRING_ENABLE

Signal slave to enable or disable corresponding vring.

SEND_RARP

Ask vhost user backend to broadcast a fake RARP to notify the migration is terminated for guest that does not support GUEST_ANNOUNCE.

NET_SET_MTU

Set host MTU value exposed to the guest.

SET_SLAVE_REQ_FD

Set the socket file descriptor for slave initiated requests.

IOTLB_MSG

Send IOTLB messages with struct vhost_iotlb_msg as payload.

SET_VRING_ENDIAN

Set the endianness of a VQ for legacy devices.

GET_CONFIG

Fetch the contents of the virtio device configuration space.

SET_CONFIG

Change the contents of the virtio device configuration space.

CREATE_CRYPTO_SESSION

Create a session for crypto operation.

CLOSE_CRYPTO_SESSION

Close a session for crypto operation.

POSTCOPY_ADVISE

Advise slave that a migration with postcopy enabled is underway.

POSTCOPY_LISTEN

Advise slave that a transition to postcopy mode has happened.

POSTCOPY_END

Advise that postcopy migration has now completed.

GET_INFLIGHT_FD

Get a shared buffer from slave.

SET_INFLIGHT_FD

Send the shared inflight buffer back to slave.

GPU_SET_SOCKET

Sets the GPU protocol socket file descriptor.

RESET_DEVICE

Ask the vhost user backend to disable all rings and reset all internal device state to the initial state.

VRING_KICK

Indicate that a buffer was added to the vring instead of signalling it using the vring’s kick file descriptor.

GET_MAX_MEM_SLOTS

Return a u64 payload containing the maximum number of memory slots.

ADD_MEM_REG

Update the memory tables by adding the region described.

REM_MEM_REG

Update the memory tables by removing the region described.

SET_STATUS

Notify the backend with updated device status as defined in the VIRTIO specification.

GET_STATUS

Query the backend for its device status as defined in the VIRTIO specification.

MAX_CMD

Upper bound of valid commands.