Trait vmm_vhost::vhost_user::VhostUserSlaveReqHandler

pub trait VhostUserSlaveReqHandler {
    fn set_owner(&self) -> Result<()>;

    fn reset_owner(&self) -> Result<()>;

    fn get_features(&self) -> Result<u64>;

    fn set_features(&self, features: u64) -> Result<()>;

    fn set_mem_table(
        &self, 
        ctx: &[VhostUserMemoryRegion], 
        files: Vec<File>
    ) -> Result<()>;

    fn set_vring_num(&self, index: u32, num: u32) -> Result<()>;

    fn set_vring_addr(
        &self, 
        index: u32, 
        flags: VhostUserVringAddrFlags, 
        descriptor: u64, 
        used: u64, 
        available: u64, 
        log: u64
    ) -> Result<()>;

    fn set_vring_base(&self, index: u32, base: u32) -> Result<()>;

    fn get_vring_base(&self, index: u32) -> Result<VhostUserVringState>;

    fn set_vring_kick(&self, index: u8, fd: Option<File>) -> Result<()>;

    fn set_vring_call(&self, index: u8, fd: Option<File>) -> Result<()>;

    fn set_vring_err(&self, index: u8, fd: Option<File>) -> Result<()>;

    fn get_protocol_features(&self) -> Result<VhostUserProtocolFeatures>;

    fn set_protocol_features(&self, features: u64) -> Result<()>;

    fn get_queue_num(&self) -> Result<u64>;

    fn set_vring_enable(&self, index: u32, enable: bool) -> Result<()>;

    fn get_config(
        &self, 
        offset: u32, 
        size: u32, 
        flags: VhostUserConfigFlags
    ) -> Result<Vec<u8>>;

    fn set_config(
        &self, 
        offset: u32, 
        buf: &[u8], 
        flags: VhostUserConfigFlags
    ) -> Result<()>;

    fn get_inflight_fd(
        &self, 
        inflight: &VhostUserInflight
    ) -> Result<(VhostUserInflight, File)>;

    fn set_inflight_fd(
        &self, 
        inflight: &VhostUserInflight, 
        file: File
    ) -> Result<()>;

    fn get_max_mem_slots(&self) -> Result<u64>;

    fn add_mem_region(
        &self, 
        region: &VhostUserSingleMemoryRegion, 
        fd: File
    ) -> Result<()>;

    fn remove_mem_region(
        &self, 
        region: &VhostUserSingleMemoryRegion
    ) -> Result<()>;

    fn set_slave_req_fd(&self, _vu_req: File) { ... }
}

Services provided to the master by the slave with interior mutability.

The VhostUserSlaveReqHandler trait defines the services provided to the master by the slave. And the VhostUserSlaveReqHandlerMut trait is a helper mirroring VhostUserSlaveReqHandler, but without interior mutability. The vhost-user specification defines a master communication channel, by which masters could request services from slaves. The VhostUserSlaveReqHandler trait defines services provided by slaves, and it's used both on the master side and slave side.

on the master side, a stub forwarder implementing VhostUserSlaveReqHandler will proxy service requests to slaves.

on the slave side, the SlaveReqHandler will forward service requests to a handler implementing VhostUserSlaveReqHandler.

The VhostUserSlaveReqHandler trait is design with interior mutability to improve performance for multi-threading.

Required methods

fn set_owner(&self) -> Result<()>

fn reset_owner(&self) -> Result<()>

fn get_features(&self) -> Result<u64>

fn set_features(&self, features: u64) -> Result<()>

fn set_mem_table(
&self,
ctx: &[VhostUserMemoryRegion],
files: Vec<File>
) -> Result<()>

fn set_vring_num(&self, index: u32, num: u32) -> Result<()>

fn set_vring_addr(
&self,
index: u32,
flags: VhostUserVringAddrFlags,
descriptor: u64,
used: u64,
available: u64,
log: u64
) -> Result<()>

fn set_vring_base(&self, index: u32, base: u32) -> Result<()>

fn get_vring_base(&self, index: u32) -> Result<VhostUserVringState>

fn set_vring_kick(&self, index: u8, fd: Option<File>) -> Result<()>

fn set_vring_call(&self, index: u8, fd: Option<File>) -> Result<()>

fn set_vring_err(&self, index: u8, fd: Option<File>) -> Result<()>

fn get_protocol_features(&self) -> Result<VhostUserProtocolFeatures>

fn set_protocol_features(&self, features: u64) -> Result<()>

fn get_queue_num(&self) -> Result<u64>

fn set_vring_enable(&self, index: u32, enable: bool) -> Result<()>

fn get_config(
&self,
offset: u32,
size: u32,
flags: VhostUserConfigFlags
) -> Result<Vec<u8>>

fn set_config(
&self,
offset: u32,
buf: &[u8],
flags: VhostUserConfigFlags
) -> Result<()>

fn get_inflight_fd(
&self,
inflight: &VhostUserInflight
) -> Result<(VhostUserInflight, File)>

fn set_inflight_fd(
&self,
inflight: &VhostUserInflight,
file: File
) -> Result<()>

fn get_max_mem_slots(&self) -> Result<u64>

fn add_mem_region(
&self,
region: &VhostUserSingleMemoryRegion,
fd: File
) -> Result<()>

fn remove_mem_region(&self, region: &VhostUserSingleMemoryRegion) -> Result<()>

Loading content...
Provided methods

fn set_slave_req_fd(&self, _vu_req: File)

Loading content...