Crate vmm_vhost

Virtio Vhost Backend Drivers

Virtio devices use virtqueues to transport data efficiently. The first generation of virtqueue is a set of three different single-producer, single-consumer ring structures designed to store generic scatter-gather I/O. The virtio specification 1.1 introduces an alternative compact virtqueue layout named “Packed Virtqueue”, which is more friendly to memory cache system and hardware implemented virtio devices. The packed virtqueue uses read-write memory, that means the memory will be both read and written by both host and guest. The new Packed Virtqueue is preferred for performance.

Vhost is a mechanism to improve performance of Virtio devices by delegate data plane operations to dedicated IO service processes. Only the configuration, I/O submission notification, and I/O completion interruption are piped through the hypervisor. It uses the same virtqueue layout as Virtio to allow Vhost devices to be mapped directly to Virtio devices. This allows a Vhost device to be accessed directly by a guest OS inside a hypervisor process with an existing Virtio (PCI) driver.

The initial vhost implementation is a part of the Linux kernel and uses ioctl interface to communicate with userspace applications. Dedicated kernel worker threads are created to handle IO requests from the guest.

Later Vhost-user protocol is introduced to complement the ioctl interface used to control the vhost implementation in the Linux kernel. It implements the control plane needed to establish virtqueues sharing with a user space process on the same host. It uses communication over a Unix domain socket to share file descriptors in the ancillary data of the message. The protocol defines 2 sides of the communication, master and slave. Master is the application that shares its virtqueues. Slave is the consumer of the virtqueues. Master and slave can be either a client (i.e. connecting) or server (listening) in the socket communication.

Modules

connection

Common data structures for listener and endpoint.

message

Define communication messages for the vhost-user protocol.

Structs

Master

Struct for the vhost-user master endpoint.

MasterReqHandler

Server to handle service requests from slaves from the slave communication channel.

SlaveFsCacheReq

Request proxy to send vhost-user-fs slave requests to the master through the slave communication channel.

SlaveListener

Vhost-user slave side connection listener.

SlaveReqHandler

Server to handle service requests from masters from the master communication channel.

SlaveReqHelper

Abstracts |Endpoint| related operations for vhost-user slave implementations.

VhostUserMemoryRegionInfo

Memory region configuration data.

VringConfigData

Vring configuration data.

Enums

Error

Errors for vhost-user operations

Protocol

Vhost-user protocol variants used for the communication.

Constants

VHOST_MAX_MEMORY_REGIONS

Maximum number of memory regions supported.

Traits

VhostBackend

An interface for setting up vhost-based backend drivers with interior mutability.

VhostBackendMut

An interface for setting up vhost-based backend drivers.

VhostUserMaster

Trait for vhost-user master to provide extra methods not covered by the VhostBackend yet.

VhostUserMasterReqHandler

Define services provided by masters for the slave communication channel.

VhostUserMasterReqHandlerMut

A helper trait mirroring VhostUserMasterReqHandler but without interior mutability.

VhostUserSlaveReqHandler

Services provided to the master by the slave with interior mutability.

VhostUserSlaveReqHandlerMut

Services provided to the master by the slave without interior mutability.

Type Definitions

HandlerResult

Result of request handler.

Result

Result of vhost-user operations