Trait vmm_vhost::connection::EndpointExt

pub trait EndpointExt<R: Req>: Endpoint<R> {
    fn send_iovec_all(
        &mut self, 
        iovs: &mut [&[u8]], 
        fds: Option<&[RawFd]>
    ) -> Result<usize> { ... }

    fn send_header(
        &mut self, 
        hdr: &VhostUserMsgHeader<R>, 
        fds: Option<&[RawFd]>
    ) -> Result<()> { ... }

    fn send_message<T: Sized + DataInit>(
        &mut self, 
        hdr: &VhostUserMsgHeader<R>, 
        body: &T, 
        fds: Option<&[RawFd]>
    ) -> Result<()> { ... }

    fn send_message_with_payload<T: Sized + DataInit>(
        &mut self, 
        hdr: &VhostUserMsgHeader<R>, 
        body: &T, 
        payload: &[u8], 
        fds: Option<&[RawFd]>
    ) -> Result<()> { ... }

    fn recv_data(&mut self, len: usize) -> Result<Vec<u8>> { ... }

    fn recv_into_bufs_all(
        &mut self, 
        bufs: &mut [&mut [u8]]
    ) -> Result<(usize, Option<Vec<File>>)> { ... }

    fn recv_header(
        &mut self
    ) -> Result<(VhostUserMsgHeader<R>, Option<Vec<File>>)> { ... }

    fn recv_body<T: Sized + DataInit + Default + VhostUserMsgValidator>(
        &mut self
    ) -> Result<(VhostUserMsgHeader<R>, T, Option<Vec<File>>)> { ... }

    fn recv_payload_into_buf<T: Sized + DataInit + Default + VhostUserMsgValidator>(
        &mut self, 
        buf: &mut [u8]
    ) -> Result<(VhostUserMsgHeader<R>, T, usize, Option<Vec<File>>)> { ... }
}

Abstracts VVU message parsing, sending and receiving.

Provided methods

fn send_iovec_all(
&mut self,
iovs: &mut [&[u8]],
fds: Option<&[RawFd]>
) -> Result<usize>

Sends all bytes from scatter-gather vectors with optional attached file descriptors. Will loop until all data has been transfered.

Return:

number of bytes sent on success

TODO

This function takes a slice of &[u8] instead of IoSlice because the internal cursor needs to be moved by advance_slices(). Once IoSlice::advance_slices() becomes stable, this should be updated. https://github.com/rust-lang/rust/issues/62726.

fn send_header(
&mut self,
hdr: &VhostUserMsgHeader<R>,
fds: Option<&[RawFd]>
) -> Result<()>

Sends a header-only message with optional attached file descriptors.

Return:

number of bytes sent on success

PartialMessage: received a partial message.

backend specific errors

fn send_message<T: Sized + DataInit>(
&mut self,
hdr: &VhostUserMsgHeader<R>,
body: &T,
fds: Option<&[RawFd]>
) -> Result<()>

Send a message with header and body. Optional file descriptors may be attached to the message.

Return:

number of bytes sent on success

OversizedMsg: message size is too big.

PartialMessage: received a partial message.

backend specific errors

fn send_message_with_payload<T: Sized + DataInit>(
&mut self,
hdr: &VhostUserMsgHeader<R>,
body: &T,
payload: &[u8],
fds: Option<&[RawFd]>
) -> Result<()>

Send a message with header, body and payload. Optional file descriptors may also be attached to the message.

Return:

number of bytes sent on success

OversizedMsg: message size is too big.

PartialMessage: received a partial message.

IncorrectFds: wrong number of attached fds.

backend specific errors

fn recv_data(&mut self, len: usize) -> Result<Vec<u8>>

Reads len bytes at most.

Return:

(number of bytes received, buf) on success

fn recv_into_bufs_all(
&mut self,
bufs: &mut [&mut [u8]]
) -> Result<(usize, Option<Vec<File>>)>

Reads all bytes into the given scatter/gather vectors with optional attached files. Will loop until all data has been transferred.

Return:

(number of bytes received, [received fds]) on success

SocketBroken: the underline socket is broken.

SocketError: other socket related errors.

TODO

This function takes a slice of &mut [u8] instead of IoSliceMut because the internal cursor needs to be moved by advance_slices_mut(). Once IoSliceMut::advance_slices() becomes stable, this should be updated. https://github.com/rust-lang/rust/issues/62726.

fn recv_header(&mut self) -> Result<(VhostUserMsgHeader<R>, Option<Vec<File>>)>

Receive a header-only message with optional attached files. Note, only the first MAX_ATTACHED_FD_ENTRIES file descriptors will be accepted and all other file descriptor will be discard silently.

Return:

(message header, [received files]) on success.

PartialMessage: received a partial message.

InvalidMessage: received a invalid message.

backend specific errors

fn recv_body<T: Sized + DataInit + Default + VhostUserMsgValidator>(
&mut self
) -> Result<(VhostUserMsgHeader<R>, T, Option<Vec<File>>)>

Receive a message with optional attached file descriptors. Note, only the first MAX_ATTACHED_FD_ENTRIES file descriptors will be accepted and all other file descriptor will be discard silently.

Return:

(message header, message body, [received files]) on success.

PartialMessage: received a partial message.

InvalidMessage: received a invalid message.

backend specific errors

fn recv_payload_into_buf<T: Sized + DataInit + Default + VhostUserMsgValidator>(
&mut self,
buf: &mut [u8]
) -> Result<(VhostUserMsgHeader<R>, T, usize, Option<Vec<File>>)>

Receive a message with optional payload and attached file descriptors. Note, only the first MAX_ATTACHED_FD_ENTRIES file descriptors will be accepted and all other file descriptor will be discard silently.

Return:

(message header, message body, size of payload, [received files]) on success.

PartialMessage: received a partial message.

InvalidMessage: received a invalid message.

backend specific errors