Struct vmm_vhost::SlaveReqHandler

pub struct SlaveReqHandler<S: VhostUserSlaveReqHandler, E: Endpoint<MasterReq>> { /* fields omitted */ }

Server to handle service requests from masters from the master communication channel.

The SlaveReqHandler acts as a server on the slave side, to handle service requests from masters on the master communication channel. It’s actually a proxy invoking the registered handler implementing VhostUserSlaveReqHandler to do the real work.

The lifetime of the SlaveReqHandler object should be the same as the underline Unix Domain Socket, so it gets simpler to recover from disconnect.

Implementations

impl<S: VhostUserSlaveReqHandler> SlaveReqHandler<S, SocketEndpoint<MasterReq>>

pub fn from_stream(socket: UnixStream, backend: Arc<S>) -> Self

Create a vhost-user slave endpoint from a connected socket.

impl<S: VhostUserSlaveReqHandler, E: Endpoint<MasterReq>> SlaveReqHandler<S, E>

pub fn connect(path: &str, backend: Arc<S>) -> Result<Self>

Create a new vhost-user slave endpoint.

Arguments

path - path of Unix domain socket listener to connect to

backend - handler for requests from the master to the slave

pub fn set_failed(&mut self, error: i32)

Mark endpoint as failed with specified error code.

pub fn handle_request(&mut self) -> Result<()>

Main entrance to server slave request from the slave communication channel.

Receive and handle one incoming request message from the master. The caller needs to:

serialize calls to this function

decide what to do when error happens

optional recover from failure