Module vmm_vhost::vhost_user

The protocol for vhost-user is based on the existing implementation of vhost for the Linux Kernel. The protocol defines two sides of the communication, master and slave. Master is the application that shares its virtqueues. Slave is the consumer of the virtqueues.

The communication channel between the master and the slave includes two sub channels. One is used to send requests from the master to the slave and optional replies from the slave to the master. This sub channel is created on master startup by connecting to the slave service endpoint. The other is used to send requests from the slave to the master and optional replies from the master to the slave. This sub channel is created by the master issuing a VHOST_USER_SET_SLAVE_REQ_FD request to the slave with an auxiliary file descriptor.

Unix domain socket is used as the underlying communication channel because the master needs to send file descriptors to the slave.

Most messages that can be sent via the Unix domain socket implementing vhost-user have an equivalent ioctl to the kernel implementation.

Modules

message

Define communication messages for the vhost-user protocol.

Structs

Listener

Unix domain socket listener for accepting incoming connections.

Master

Struct for the vhost-user master endpoint.

MasterReqHandler

Server to handle service requests from slaves from the slave communication channel.

SlaveFsCacheReq

Request proxy to send vhost-user-fs slave requests to the master through the slave communication channel.

SlaveListener

Vhost-user slave side connection listener.

SlaveReqHandler

Server to handle service requests from masters from the master communication channel.

Enums

Error

Errors for vhost-user operations

Traits

VhostUserMaster

Trait for vhost-user master to provide extra methods not covered by the VhostBackend yet.

VhostUserMasterReqHandler

Define services provided by masters for the slave communication channel.

VhostUserMasterReqHandlerMut

A helper trait mirroring VhostUserMasterReqHandler but without interior mutability.

VhostUserSlaveReqHandler

Services provided to the master by the slave with interior mutability.

VhostUserSlaveReqHandlerMut

Services provided to the master by the slave without interior mutability.

Type Definitions

HandlerResult

Result of request handler.

Result

Result of vhost-user operations